# Refactor `find` in terms of clean `Record` API