@sholderbach 2023-11-08
Performance measured with 

```nu
1..1000 | each { scope commands | timeit { find 42 } } | math avg
```

as `scope commands` returns a reasonably large known table with a decent number of columns.

## Before (base commit)

`14ms 215µs 420ns`

## With this 

`13ms 329µs 18ns`